+++ src/main/java/me/pajic/toolpouch/platform/Platform.java
@@ -4,7 +4,6 @@ import me.pajic.toolpouch.menu.ToolPouchMenu;
 import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.tags.TagKey;
-import net.minecraft.world.InteractionResult;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.inventory.MenuType;
 import net.minecraft.world.item.ItemStack;
@@ -25,7 +24,7 @@ public interface Platform {
 
 	MenuType<ToolPouchMenu> constructMenu();
 
-	InteractionResult openToolPouchScreen(Player player, ItemStack backpack);
+	void openToolPouchScreen(Player player, ItemStack backpack);
 
 	String getTagTranslationKey(TagKey<?> tagKey);
 }